module lavalinkext
