+++ extension/src/index.ts
@@ -1,3 +1,19 @@
-﻿export { voiceExt } from "./voiceExt";
-export { lyricsExt } from "./lyricsExt";
+﻿// Main extension
 export { lavalinkExt } from "./lavalinkExt";
+
+// Types
+export * from "./types/lavalink";
+
+// Managers
+export { NodeManager } from "./managers/NodeManager";
+export { PlayerStateManager } from "./managers/PlayerStateManager";
+
+// Handlers
+export { WebSocketHandler } from "./handlers/WebSocketHandler";
+export { VoiceHandler } from "./handlers/VoiceHandler";
+
+// Resolvers
+export { TrackResolver } from "./resolvers/TrackResolver";
+
+// Utils
+export * from "./utils/helpers";
